gadget, snap: move gadget validation to separate 'gadget' package
